+++ CHANGELOG.md
@@ -15,7 +15,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add class for non-modulating heat pumps (issue #476).
 - Added default secondary temperature to EER and COP class (issue #476).
 - Added _get_min_power to _Efficiency class (issue #476).
--
 
 ### Changed
 
@@ -23,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Change implementation ConicalProbe for speed improvement (issue #472).
 - Change implementation of interpolation for _Efficiency class to cope with non-gridded data input (issue #472).
 - Change implementation of combine n heat pumps (issue #472).
+- Change implementation for conversion volume to weight percentage (issue #481).
 
 ### Fixed